requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
ebooklib>=0.18
beautifulsoup4>=4.12.0
google-generativeai>=0.8.0
edge-tts>=6.1.0
pydub>=0.25.0
